refactor: 更新记分板相关文件路径
- 新建文件夹 `core/initialization`
- 将 `scoreBase` 目录从 `lib/xboyPackage` 移动到 `core/initialization`
- 更新相关文件的引用路径
