+++ src/core/initialization/scoreBase/rw.ts
