+++ src/core/initialization/scoreBase/verifyDataBase.ts
